requirements.txt
accelerate==0.29.3
certifi==2024.2.2
charset-normalizer==3.3.2
diffusers==0.27.2
diskcache==5.6.3
filelock==3.13.4
fsspec==2024.3.1
huggingface_hub==0.22.2
idna==3.7
importlib_metadata==7.1.0
Jinja2==3.1.3
llama_cpp_python==0.2.65
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
packaging==24.0
pillow==10.2.0
psutil==5.9.8
PyYAML==6.0.1
regex==2024.4.16
requests==2.31.0
safetensors==0.4.3
schedule==1.2.1
sympy==1.12
tokenizers==0.19.1
torch==2.3.0
torchaudio==2.3.0
torchvision==0.18.0
tqdm==4.66.2
transformers==4.40.1
typing_extensions==4.11.0
urllib3==2.2.1
zipp==3.18.1
